+++ components/Toast.tsx
@@ -42,6 +42,8 @@ const Toast = () => {
     return () => {
       if (showTimerId) clearTimeout(showTimerId);
       if (animateTimerId) clearTimeout(animateTimerId);
+
+      useToastStore.setState({ toastId: 0 });
     };
   }, [showTimerId, animateTimerId]);
